smart_irrigation/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.1.0

# Database and API
supabase>=1.0.3
python-dotenv>=0.19.0

# Scheduling and timing
schedule>=1.1.0

# Data processing and visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utility packages
tqdm>=4.62.0  # For progress bars
pytz>=2021.1  # For timezone handling

# Development tools
pytest>=6.2.5  # For testing
black>=21.7b0  # For code formatting
flake8>=3.9.0  # For code linting